requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ollama==0.6.1
openai==2.15.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5

pandas
plotly
streamlit
pillow
rembg